Removed unnecessary data duplication, necessary to pass things to GPU.

+++ src/decentralizepy/datasets/MovieLens.py
@@ -290,8 +290,10 @@ class MatrixFactorization(Model):
         """
         Forward pass of the model, it does matrix multiplication and returns predictions for given users and items.
         """
-        users = torch.LongTensor(data[:, 0]) - 1
-        items = torch.LongTensor(data[:, 1]) - 1
+        users = data[:, 0] - 1
+        items = data[:, 1] - 1
+        # logging.debug("User dtype: %s", users.dtype)
+        # logging.debug("items dtype: %s", items.dtype)
         u, it = self.user_factors(users), self.item_factors(items)
         x = (u * it).sum(dim=1, keepdim=True)
         return x.squeeze(1)
